Adding new CTPM sepolicy and domain.
Right now the ConnectivityThermalPowerManager is part of the
platform_app sepolicy domain. platform_app's don't have access to the
power_stats_hal. So following the security principle of least privilege
we need to define our own domain with the permissions we previously had
as part of the old domain.

Test: Compiled and loaded onto device and saw no "avc: denied" sepolicy
errors
Bug: 202028512
Ignore-AOSP-First: CTPM is a pixel only feature for thermal / power
mitigation

Change-Id: I140b704fc39959ba683c09ebe9ea65f7bf94452b